refactor: Phase 0 — KlineRow type unification, StorageEngine/WriteBuffer/DataIngestor fixes
- T0.1: Unify to event::KlineRow (int32_t×10000 fixed-point, per design)
- T0.2: Fix DataInitializer store_kline_freq_batch → store_kline_batch
- T0.3: Restore StorageEngine::set_write_buffer()/close() methods
- T0.4: WriteBuffer do_flush_locked() → co_await co_store_kline_batch
- T0.5: DataIngestor switches from TimeSeriesStore to StorageEngine
- T2.1: StrategyWatcher watches /quant/backtest/task/ prefix
- T2.2: StrategyRegistry adds find_by_id(uint64_t)
- T2.3: service_main integrates EtcdClient + StrategyWatcher
- Fix AffinitySharedMutex await_ready() — use CAS instead of load (was not claiming lock on fast path)
- ColumnBlock/TimeSeriesCache/WAL/WriteBuffer tests: 19/19 passing

+++ cpp/test/e2e_integration_test.cc
@@ -102,7 +102,7 @@ protected:
             row.close_price = static_cast<int32_t>(std::round(price * 10000.0));
             row.volume      = 1'000'000 + i * 10'000;
             row.amount      = static_cast<int64_t>(std::round(price * 10000.0)) * 100;
-            row.vwap        = static_cast<int64_t>(std::round(price * 10000.0));
+            row.vwap        = static_cast<int32_t>(std::round(price * 10000.0));
             rows.push_back(row);
         }
 
@@ -209,7 +209,7 @@ protected:
             row.close_price = static_cast<int32_t>(std::round(price * 10000.0));
             row.volume      = 1'000'000 + i * 10'000;
             row.amount      = static_cast<int64_t>(std::round(price * 10000.0)) * 100;
-            row.vwap        = static_cast<int64_t>(std::round(price * 10000.0));
+            row.vwap        = static_cast<int32_t>(std::round(price * 10000.0));
             rows.push_back(row);
         }
